+++ src/components/molecules/LayoutWithBottomButtons/index.tsx
@@ -1,10 +1,10 @@
 import React, {ReactElement} from 'react';
 import {StyleSheet, ViewProps, View} from 'react-native';
-import BaseButton from 'atoms/BaseButton';
 import {palette} from 'theme/palette';
 import {moderateScale, scaledForDevice} from 'scale';
 import {validVariants, parseButtonsStyles, buttonWrapperVariantStyles} from './utils';
 import type {IlayoutButtons} from './utils';
+import Button from 'molecules/Button';
 
 interface LayoutWithBottomButtonsProps extends ViewProps {
 	children: ReactElement | string;
@@ -35,6 +35,8 @@ const LayoutWithBottomButtons = ({
 		return null;
 	}
 
+	const isRoundedOnTopVariant = selectedVariant === validVariants.roundedOnTop;
+
 	const styles = StyleSheet.create({
 		Container: {
 			flex: 1,
@@ -54,7 +56,7 @@ const LayoutWithBottomButtons = ({
 		},
 		ButtonsWrapper: {
 			...buttonWrapperVariantStyles(selectedVariant),
-			flexDirection: 'row',
+			flexDirection: isRoundedOnTopVariant ? 'column' : 'row',
 			justifyContent: 'space-between',
 			backgroundColor: validBtnBgColor,
 			marginTop: 'auto',
@@ -72,10 +74,10 @@ const LayoutWithBottomButtons = ({
 			<View style={[styles.Container]} {...props}>
 				{children}
 				<View style={[styles.FullButtonsRoundedWrapper]}>
-					<BaseButton {...fullWidthButton} />
+					<Button {...fullWidthButton} />
 					<View style={[styles.TwoButtonsWrapper]}>
 						{newBtns.map(({...buttonData}, index) => (
-							<BaseButton key={index.toString()} {...buttonData} />
+							<Button key={index.toString()} {...buttonData} />
 						))}
 					</View>
 				</View>
@@ -92,7 +94,7 @@ const LayoutWithBottomButtons = ({
 			{children}
 			<View style={[styles.ButtonsWrapper]}>
 				{parsedButtons.map(({...buttonData}, index) => (
-					<BaseButton key={index.toString()} {...buttonData} />
+					<Button key={index.toString()} {...buttonData} />
 				))}
 			</View>
 		</View>